requirements.txt
django>=4.1.0,<4.2.0
psycopg2>=2.8
djangorestframework
djangorestframework-simplejwt
pillow